gowithtests/struct_interface_test.go
package main

import "testing"
import "math"

func testPerimeter(t *testing.T) {
	rectangle := Rectangle{10.0, 10.0}
	got := Perimeter(rectangle)
	want := 40.0

	if got != want {
		t.Errorf("got %.2f want %.2f", got, want)
	}
}

func testArea(t *testing.T) {

	checkArea := func(t testing.TB, shape Shape, want float64) {
		t.Helper()
		got := shape.Area()

		if got != want {
			t.Errorf("want %g got %g", want, got)
		}

	}

	t.Run("rectangle", func(t *testing.T) {
		rectangle := Rectangle{10.0, 10.0}
		checkArea(t, rectangle, 100.0)
	})
	t.Run("circle", func(t *testing.T) {
		circle := Circle{20.0}

		checkArea(t, circle, math.Pi*circle.radius)

	})

}
